Main

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2021/22.

Notas da Tarefa5

As principais dificuldades encontradas nesta tarefa devem-se ao facto de os exemplos dados pela equipa docente terem sido muito minimalistas, não abrangendo muitas das funcionalidades da biblioteca Gloss. Deste modo, foi necessária uma procura de informação mais profunda e foram usados métodos, algoritmos e ideias presentes nos vídeos de apoio disponibilizados pelo Cesium para a elaboração desta tarefa. Sendo uma Tarefa gráfica, a implementação de testes para a mesma não se justificou, tema abordado e esclarecido com a docente do nosso turno prático.

Fontes das imagens usadas:

imagem do jogador: https://www.kindpng.com/imgv/TJmhwJh_2d-game-character-png-transparent-png/

imagem da porta: https://pt.dreamstime.com/estilo-de-desenho-animado-medieval-vetorial-da-porta-do-castelo-isolado-em-branco-dobradi%C3%A7as-antigas-portas-e-pedras-s%C3%B3lidas-image161353249

imagem do chão: https://www.mobly.com.br/papel-de-parede-tijolos-2d-natural-464727.html

imagem da caixa: https://opengameart.org/content/top-down-2d-metal-box

Documentation

data OpcoesMenu

Foi criado o tipo de dados OpcoesMenu com o objetivo de representar e desenhar as opções do menu principal do jogo.

Constructors

Jogar
O construtor Jogar é o que permite jogar o jogo.

Sair
O construtor Sair é o que permite sair do menu.

data Nivel

Adicionalmente criou-se o tipo de dados Nivel para determinar o nível selecionado para jogar.

Constructors

Um
Construtor que se refere ao nível 1

Dois
Construtor que se refere ao nível 2

Tres
Construtor que se refere ao nível 3

data Menu

Criou-se o tipo de dados Menu no qual estão representados todos as possibilidades de estado do jogo de modo a ser possível desenhar todo o jogo.

Constructors

Controlar OpcoesMenu
O construtor Controlar vai receber uma opção do menu para poder interagir com o mesmo.

EscolheNivel Nivel
O construtor EscolheNivel recebe um nível para o poder inicializar.

ModoJogo Nivel
O construtor ModoJogo recebe o nível escolhido para desenhar esse mesmo nível.

VenceuJogo
O construtor VenceuJogo serve para determinar quando o jogador concluiu o nível com sucesso.

type Mundo = (Menu, Jogo, Coordenadas)

O tipo de dados Mundo é um tuplo que armazena a informação do estado do menu na primeira componente, o jogo a ser inicializado na segunda componente e a terceira componente guarda as coordenadas da Porta do jogo de modo a futuramente se saber quando o jogo foi concluido com sucesso.

type EstadoGloss = (Mundo, [(Peca, Picture)], [(Direcao, Picture)])

Por fim, o tipo de dados EstadoGloss é o mais global, que armazena:

O mundo;

Uma lista de pares (Peca,Picture) em que a primeira componente corresponde a um tipo de peça e a segunda à imagem correspondente;

Uma lista de pares (Direcao,Picture) em que a primeira componente corresponde à direção do jogador e a segunda à imagem correspondente.

desenha :: EstadoGloss -> Picture

A função desenha é uma das principais funções desta Tarefa, sendo ela responsável por desenhar todo o jogo, tendo em conta o tipo de Menu presente no estado a ser desenhado.

desenhaOpcoesMenu :: String -> Picture

A função desenhaOpcoesMenu é uma função auxiliar da função desenha que recebe a string s e converte-a numa Picture com o texto contido nessa string, de modo a apresentar o texto presente nos menus.

moverTecla :: Event -> EstadoGloss -> EstadoGloss

A função moverTecla é uma das funções principais e é responsável pela atualização do estado do jogo tendo em conta a tecla premida pelo jogador e o estado imediatamente anterior quando a tecla foi premida.

getMapa :: Mundo -> Mapa

A função getMapa é uma função auxiliar da função converteJogo que recebe um estado e devolve o mapa contido nesse estado para depois converter as peças desse mapa numa imagem.

lP :: Float

A constante lP guarda o valor da largura das imagens que vão ser criadas para os mapas dos níveis 1 e 2.

lG :: Float

A constante lG guarda o valor da largura das imagens que vão ser criadas para o mapa do nível 3.

alturaP :: Float

A constante alturaP guarda o valor da ordenada onde as imagens vão começar a ser colocadas para os mapas dos níveis 1 e 2.

alturaG :: Float

A constante alturaG guarda o valor da ordenada onde as imagens vão começar a ser colocadas para o mapa do nível 3.

comprimentoP :: Float

A constante comprimentoP guarda o valor da abcissa onde as imagens vão começar a ser colocadas para os mapas dos níveis 1 e 2.

comprimentoG :: Float

A constante comprimentoG guarda o valor da abcissa onde as imagens vão começar a ser colocadas para o mapa do nível 3.

escJogP :: Float

A constante escJogP guarda o valor da escala para a imagem que representa o jogador para os mapas dos níveis 1 e 2.

escJogG :: Float

A constante escJogG guarda o valor da escala para a imagem que representa o jogador para o mapa do nível 3.

escPecP :: Float

A constante escPecP guarda o valor da escala para as imagens que vão substituir as peças dos mapas dos níveis 1 e 2.

escPecG :: Float

A constante escPecG guarda o valor da escala para as imagens que vão substituir as peças do mapa do nível 3.

vazioPic :: Picture

A imagem vazioPic é um simples quadrado branco que vai representar as peças Vazio do mapa

desenhaJogador :: Jogador -> Float -> Float -> Float -> Float -> [(Peca, Picture)] -> Float -> [(Direcao, Picture)] -> [Picture]

A função desenhaJogador é a função que recebe um Jogador e converte-o numa lista de imagens, colocando-as nas posições corretas, convertendo as coordenadas do Jogador. Para tal recebe os valores de comprimento, altura, largura, escala das peças e escala dos jogadores para definir o tamanho com o qual a sua imagem vai ser desenhada. Recebe também as listas que fazem o mapeamento entre as peças e as respetivas imagens e a direção do jogador com as respetivas imagens. O resultado é uma lista de Pictures pois caso o booleano do jogador que determina se o Jogador carrega caixa for True significa que se tem que adicionar uma imagem de uma caixa por exatamente por cima do Jogador.

desenhaPeca :: Float -> Float -> Float -> Peca -> Float -> [(Peca, Picture)] -> Picture

A função desenhaPeca é a função que recebe uma peça, a posição onde esta vai ser colocada, a lista onde está mapeada a imagem correspondente à peça e desenha a imagem com a escala correspondente.

desenhaLinha :: Float -> Float -> Float -> [Peca] -> Float -> [(Peca, Picture)] -> [Picture]

A função desenhaLinha é a função que vai desenhar uma linha do mapa, desenhando a imagem da primeira peça da linha através da função desenhaPeca e recursivamente desenha o resto da linha, atualizando a posição onde a imagem seguinte vai ser colocada.

desenhaMapa :: Float -> Float -> Float -> Mapa -> Float -> [(Peca, Picture)] -> [Picture]

A função desenhaMapa é a função que vai converter um mapa na respetiva representação por imagens, convertendo a primeira linha e recursivamente desenha o resto das linhas, atualizando a posição onde a linha seguinte vai ser colocada.

converteJogo :: EstadoGloss -> Picture

A função converteJogo é a função que vai converter o estado numa imagem, de modo a representar gráficamente os diferentes níveis do jogo, usando as constantes de tamanhos e escalas referentes ao nível selecionado.

getJogador :: Mundo -> Jogador

A função getJogador é uma função auxiliar da função converteJogo que recebe um estado e devolve o jogador contido nesse estado para converter o jogador desse estado na respetiva imagem.

window :: Display

Definição do gloss para que o jogo seja corrido em fullscreen.

fr :: Int

Definição do gloss para que define o frame rate do jogo.

tempo :: Float -> EstadoGloss -> EstadoGloss

Definição do gloss que altera o estado do jogo consoante a passagem do tempo. Neste caso não ocorre nenhuma atualização.

background :: Color

Definição do gloss que define a cor do fundo da janela onde corre o jogo.

constroiEstadoGloss :: Mundo -> [(Peca, Picture)] -> [(Direcao, Picture)] -> EstadoGloss

A função constroiEstadoGloss é uma função usada pela função main que vai construir o estado inicial quando o jogo começa a ser executado.

estado1 :: Mundo

O estado1 é o estado que é carregado quando no menu é selecionado o nível 1.

m1 :: Mapa

m1 é o mapa do nível 1.

j1 :: Jogador

j1 são as definições iniciais do jogador do nível 1.

cp1 :: Coordenadas

cp1 são as coordenadas da porta do mapa do nível 1.

estado2 :: Mundo

O estado2 é o estado que é carregado quando no menu é selecionado o nível 2.

m2 :: Mapa

m2 é o mapa do nível 2.

j2 :: Jogador

j2 são as definições iniciais do jogador do nível 2.

cp2 :: Coordenadas

cp2 são as coordenadas da porta do mapa do nível 2.

estado3 :: Mundo

O estado3 é o estado que é carregado quando no menu é selecionado o nível 3.

m3 :: Mapa

m3 é o mapa do nível 3.

j3 :: Jogador

j3 são as definições iniciais do jogador do nível 3.

cp3 :: Coordenadas

cp3 são as coordenadas da porta do mapa do nível 3.

estadoInicial :: Mundo

O estadoInicial é o estado usado para inicializar o gloss no modo Menu Controlar Jogar. O resto dos parâmetros são meramente aleatórios e serão substituídos na escolha do nível.

main :: IO ()

Função principal do módulo, que faz o jogo ser executado.