Tarefa1_2021li1g066

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

maiorYmapa :: [(Peca, Coordenadas)] -> Int

A função maiorYmapa calcula o número de linhas da matriz recorrendo ao método de descobrir o maior y da lista de peças.

linhasM :: [(Peca, Coordenadas)] -> Int

A função linhasM calcula o número de linhas da matriz somando 1 à maior ordenada da lista de peças dada. Isto deve-se a existirem coordenadas com o valor 0.

maiorXmapa :: [(Peca, Coordenadas)] -> Int

A função maiorXmapa calcula o número de colunas da matriz, recorrendo ao método de descobrir o maior x da lista de peças.

colunasM :: [(Peca, Coordenadas)] -> Int

A função colunasM calcula o número de colunas da matriz somando 1 à maior abcissa da lista de peças dada.

getPeca :: Coordenadas -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função getPeca é uma função auxiliar muito útil que recebe umas coordenadas (x,y) e uma lista de pecas e seleciona e retorna a peça com essas coordenadas. A peça final é colocada numa lista para efeitos de recursividade.

getPecasAdj :: (Peca, Coordenadas) -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função getPecasAdj junta numa lista as peças acima, abaixo e à direita da peça dada. Serve para mais adiante se verificar se a peça dada está junta a uma peca Bloco, determinando a continuidade do chão.

procUltPecasColunas :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função procUltPecasColunas junta numa lista as últimas peças de cada coluna. Serve para futuramente se verificar a validade do chão.

filtraColunaX :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função filtraColunaX é a função auxiliar da função procUltPecasColunas, que dado um inteiro n vai filtrar a lista de peças dada e retornar apenas as peças que pertençam à coluna n.

verPecasAdj :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [[(Peca, Coordenadas)]]

A função verPecasAdj pega na lista de últimas peças das colunas e no mapa e constrói uma lista com a lista de peças adjacentes respetivamente.

checkTamanhos :: [[(Peca, Coordenadas)]] -> [Bool]

A função checkTamanhos verifica se o tamanho da lista de peças adjacentes é pelo menos 1, para assegurar que todas as últimas peças Bloco estão ligadas a outra peça Bloco.

chaoOk :: [(Peca, Coordenadas)] -> Bool

A função chaoOk é uma auxiliar da função principal que valida o chão do mapa. Para tal é necessário que haja continuidade das peças Bloco. Para resolver este problema seguimos a seguinte estratégia :

Juntar numa lista as últimas peças de cada coluna do mapa, da esquerda para a direita

Verificar quais as peças que estão acima, abaixo e à direita de cada peça da lista do ponto anterior. Para haver continuidade no chão, os blocos têm que estar ligados por uma dessas posições.

Para que haja continuidade, tem que haver pelo menos uma peça adjacente, por isso o tamanho da lista de peças adjacentes tem que ser maior que 1

Verificar se todas as peças finais da coluna têm uma ou mais peça adjacente

No fim, fazer a conjunção dos resultados, validando ou não o chão do mapa

contaPorta :: [(Peca, Coordenadas)] -> Int

A função contaPorta conta quantas portas existem no mapa. Para validação tem de ser contada apenas uma porta

portasOk :: [(Peca, Coordenadas)] -> Bool

A função portasOk é uma auxiliar da função principal que confirma a validade do número de portas. Se apenas existir 1 retorna True, caso contrário retorna False.

comparaPecas :: (Peca, Coordenadas) -> [(Peca, Coordenadas)] -> Bool

A função comparaPecas recebe uma peça e verifica se existe alguma com coordenadas iguais numa lista de peças. Havendo coordenadas iguais não é valido e retorna False. Caso contrário é válido e retorna True.

pecasOk :: [(Peca, Coordenadas)] -> Bool

A função pecasOk é uma auxiliar da função principal. Verifica se exitem peças com coordenadas iguais. Caso existam, não é valido e retorna False. Caso contrário é válido e retorna True.

getPecaInferior :: (Peca, Coordenadas) -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função getPecaInferior retorna a peça inferior à peça dada. O resultado é inserido numa lista para ser possível retornar um resultado no caso de paragem da recursividade e também para futuramente poder ser comparável.

caixasOk :: [(Peca, Coordenadas)] -> Bool

A função caixasOk é uma função auxiliar da principal. Verifica se as caixas estão bem posicionadas, ou seja, se estão em cima de outra caixa ou bloco. Para tal foi necessário separar as caixas do resto das peças, verificar a validade de cada uma, juntando os respetivos resultados numa lista e fazendo a conjunção desses resultados.

juntaCaixas :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função juntaCaixas junta todas as caixas de uma lista de peças numa nova lista para mais tarde verificar as respetivas peças inferiores.

validadeCaixas :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [Bool]

A função validadeCaixas verifica a validade das caixas e utiliza a função verificaPecaInferior, guardando os respetivos resultados numa lista, para mais tarde se fazer a conjunção desses resultados.

verificaPecaInferior :: (Peca, Coordenadas) -> [(Peca, Coordenadas)] -> Bool

A função verificaPecaInferior verifica a validade da peça inferior de uma caixa. É necessária para a validação das caixas.

contaVazio :: [(Peca, Coordenadas)] -> Int

A função contaVazio é auxiliar da função vazioOk que conta o número de peças Vazio. Para tal, calcula o número total de peças e retira o número de peças descritas na lista de peças. O resultado será o número de peças vazias.

vazioOk :: [(Peca, Coordenadas)] -> Bool

A função vazioOk é uma das principais funções auxiliares da função principal desta tarefa. Verifica se existe pelo menos uma peça Vazio na lista de peças.

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

A função validaPotencialMapa é a função principal desta tarefa. Consiste em verificar se o mapa l corresponde a uma mapa válido, obedecendo às 5 condições do enunciado.

Exemplo de utilização:

>>> validaPotencialMapa [(Porta, (0, 3)),(Bloco, (0, 4)),(Bloco, (1, 4)),
                         (Bloco, (2, 4)),(Bloco, (3, 4)),(Bloco, (4, 4)),
                         (Caixa, (4, 3)),(Bloco, (5, 4)),(Bloco, (6, 4)),
                         (Bloco, (6, 3)),(Bloco, (6, 2)),(Bloco, (6, 1))]
True