Tarefa2_2021li1g066

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

geraPecasVazio :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função geraPecasVazio gera uma lista de peças Vazio com todas as coordenadas possiveis, isto é, procura na lista de peças dada as maiores componentes x e y e combina-as.

removeVazioCoordRep :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função removeVazioCoordRep pega na lista gerada na função geraPecasVazio e remove as peças dessa lista cujas coordenadas já estejam declaradas na lista de peças dada.

juntaVazioLP :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função juntaVazioLP junta à lista de peças dada a lista de peças Vazio gerada, já sem a presença de peças com coordenadas repetidas.

separaMatriz :: [(Peca, Coordenadas)] -> Int -> [[(Peca, Coordenadas)]]

Depois de se inserir as peças Vazio e as respetivas coordenadas, a função separaMatriz separa as peças nas respetivas linhas, sendo que a primeira linha da matriz corresponde à primeira lista contida na lista de resultados. Esta função recebe um valor inteiro x para separar as peças da linha x e recursivamente vai se separando incrementando 1 ao x.

separaEmLinhas :: [(Peca, Coordenadas)] -> Int -> [(Peca, Coordenadas)]

A função separaEmLinhas seleciona as peças da lista dada que contenham a ordenada igual a x, ou seja, todas as que pertençam a essa linha.

ordenaPecasLinha :: [[(Peca, Coordenadas)]] -> [[(Peca, Coordenadas)]]

A função ordenaPecasLinha ordena as peças de cada linha da lista de peças que foi anteriormente separada em linhas. Deste modo, cada peça fica no exato local da matriz correspondente às suas coordenadas.

insere :: (Peca, Coordenadas) -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função insere é uma função auxiliar que insere uma peça numa lista de peças já ordenada pelas abcissas

ordena :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função ordena é uma função auxiliar que insere a primeira peça da lista na restante parte da lista ordenada recursivamente.

removeCoordenadasLinha :: [(Peca, Coordenadas)] -> [Peca]

A função auxiliar removeCoordenadasLinha faz desaparecer as coordenadas relativas às peças de uma linha da matriz, pois estas já se encontram nas suas posições corretas.

semCoordenadas :: [[(Peca, Coordenadas)]] -> Mapa

Uma vez que as peças já estão nos seus respetivos lugares, é preciso remover as coordenadas das peças. A função semCoordenadas trata exatamente de ir removendo linha a linha as coordenadas das peças.

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

A função constroiMapa é uma das funções principais desta tarefa. Consiste em construir um Mapa através de uma lista de pares (peça, coordenada). Para tal recebe uma lista de peças e segue a seguinte estratégia:

Junta peças Vazio que não estão declaradas à lista dada;

Separa a lista nas respetivas linhas, recebendo o 0 como valor inicializador pois a primeira linha encontra-se na posiçao 0;

Ordena as peças de cada linha por ordem crescente das abcissas;

Retira as coordenadas, ficando exatamente com um Mapa.

Exemplo de utilização:

>>> constroiMapa [(Porta, (0,2)), (Bloco, (0,3)), (Bloco, (1,3)),
                  (Bloco, (2,3)), (Bloco, (3,3)), (Caixa, (4,2)),
                  (Bloco, (4,3)), (Bloco, (5,3)), (Bloco, (6,0)),
                  (Bloco, (6,1)), (Bloco, (6,2)), (Bloco, (6,3))]
[[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco],
 [Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco],
 [Porta,Vazio,Vazio,Vazio,Caixa,Vazio,Bloco],
 [Bloco,Bloco,Bloco,Bloco,Bloco,Bloco,Bloco]]

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

A função desconstroiMapa é uma das funções principais desta tarefa. Consiste em desconstruir um Mapa retornando uma lista de pares (peça, coordenada). Para tal recebe um mapa segue a seguinte estratégia:

Insere as coordenadas a cada uma das peças, inicializando com as coordenadas (0,0);

Remove as declarações de peças Vazio e das respetivas coordenadas, sendo o resultado final o objetivo desta função.

Exemplo de utilização:

>>> desconstroiMapa [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                     [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                     [Porta, Vazio, Vazio, Vazio, Caixa, Vazio, Bloco],
                     [Bloco, Bloco, Bloco, Bloco, Bloco, Bloco, Bloco]]
[(Bloco,(6,0)),(Bloco,(6,1)),(Porta,(0,2)),
 (Caixa,(4,2)),(Bloco,(6,2)),(Bloco,(0,3)),
 (Bloco,(1,3)),(Bloco,(2,3)),(Bloco,(3,3)),
 (Bloco,(4,3)),(Bloco,(5,3)),(Bloco,(6,3))]

insereCoordenadas :: Mapa -> Int -> Int -> [(Peca, Coordenadas)]

insereCoordenadasLinha :: [Peca] -> Int -> Int -> [(Peca, Coordenadas)]

eliminaVazioLista :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]