Tarefa3_2021li1g066

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2021/22.

Documentation

converteMapa :: Mapa -> String

A função converteMapa é uma das funções principais para definir a instância Show e consiste na troca das peças do mapa pelo seu carater correspondente definido anteriormente. Para tal, seguiu-se a seguinte estratégia:

Converter as peças da primeira linha do mapa e concatenar os resultados.

Caso exista linha seguinte adiciona-se um carater n no fim das peças da primeira linha

Constrói-se o resto das linhas do mapa recursivamente, concatenando-se todos os resultados para o resultado final ser uma única String

insereJogador :: Jogador -> String -> String

A função insereJogador é outra das funções principais desta Tarefa. Recebe um jogador e um mapa já convertido em String e insere o jogador nessa string na posição correspondente às suas coordenadas. Seguiu-se a seguinte estratégia:

Separar o mapa em string numa lista com as strings de cada linha;

Trocar o carater da string na posição das coordenadas do jogador pelo carater correspondente; ao jogador, tendo em conta a direção do mesmo;

Ter em conta aa componente tc do jogador (se transporta caixa ou não) e decidir se insere um carater de caixa por cima dele;

Voltar a concatenar as String numa só.

averiguaDirecao :: Direcao -> Char

A função averiguaDirecao é uma função auxiliar que atribui o carater correspondente à direção do jogador dada.

insereEnters :: [String] -> String

A função insereEnters é a função que vai concatenar a lista de strings na string final, separando as linhas com a string que contém o carater n.

inverteMapaStrToLines :: String -> [String]

A função inverteMapaStrToLines vai receber a string criada na função converteMapa e separa-a nas linhas usando a função lines.

trocaElemMapa :: Char -> Coordenadas -> [String] -> [String]

A função trocaElemMapa é a função que vai colocar o carater do jogador na linha e na coluna correspondente, tendo em conta tanto a direção do jogador e as suas coordenadas.

trocaElemLinha :: Int -> Char -> String -> String

A função trocaElemLinha é a função auxiliar da função trocaElemMapa, e serve para trocar o elemento na posição x pelo carater ch correspondente ao jogador

Orphan instances

Show Jogo
Esta é a intância show principal desta Tarefa. Consiste na representação em string de um Jogo. Para tal:

Converte as peças do mapa m numa string com os carateres correspondentes a cada peça.

Insere o carater do jogador jog na posição da string correspondente às suas coordenadas.

Exemplo de utilização:

>>> show (Jogo [ [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio],
                 [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                 [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                 [Porta, Vazio, Vazio, Vazio, Caixa, Vazio, Bloco],
                 [Bloco, Bloco, Bloco, Bloco, Bloco, Bloco, Bloco]]
                 (Jogador (6,0) Oeste False) )
"      <\n      X\n      X\nP   C X\nXXXXXXX"

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS

Show Peca
Instância que define o carater de cada tipo de peça

Instance details

Methods

showsPrec :: Int -> Peca -> ShowS

show :: Peca -> String

showList :: [Peca] -> ShowS