Tarefa4_2021li1g066

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

getPecaMapa :: Coordenadas -> Mapa -> Peca

A função getPecaMapa é uma função auxiliar muito útil nesta tarefa, que recebe umas coordenadas e seleciona a peça do mapa m correspondente.

temObstaculo :: Jogo -> Movimento -> Bool

A funçãp temObstaculo é uma função muito importante nesta tarefa. Consiste em averiguar se existe um obstáculo no mapa do jogo j para que o movimento mov possa ou não acontecer. Se o resultado for True significa que existe um obstáculo ao movimento, não podendo este ser efetuado.

obsAndarDir :: Jogo -> Bool

A função obsAndarDir é a função que diz se existe um obstáculo que impeça o movimento do jogador para a direita, estando ou não estando este a carregar uma caixa. Uma condição importante é quando o jogador já se encontra na última coluna do mapa, não podendo avançar para a direita, expressa na expressão length (head mapa) .

obsAndarEsq :: Jogo -> Bool

A função obsAndarEsq é a função que diz se existe um obstáculo que impeça o movimento do jogador para a esquerda, estando ou não estando este a carregar uma caixa. Uma condição importante é quando o jogador já se encontra na primeira coluna do mapa, não podendo avançar para a esquerda, expressa na expressão x == 0 .

obsInteragirCaixa :: Jogo -> Bool

A função obsInteragirCaixa é a função que diz se existe um obstáculo que impeça a interação do jogador com uma caixa, estando ou não estando este a carregar uma caixa. Uma condição importante é que o jogador tem que estar virado para a caixa para poder pegar nela. Caso o jogador já transporte uma caixa, tem que ver se há algum obstáculo para a deixar cair.

obsTrepar :: Jogo -> Bool

A função obsTrepar é a função que diz se existe um obstáculo que impeça a ação de trepar caixas e blocos, estando ou não estando este a carregar uma caixa. Caso o jogador esteja a transportar uma caixa, tem que ver há espaço para si e para a caixa de modo a poder subir.

moveEsq :: Jogo -> Jogador

A função moveEsq é a função que faz o jogador se mover para a esquerda caso não haja nenhum obstáculo que impeça este movimento.

moveDir :: Jogo -> Jogador

A função moveDir é a função que faz o jogador se mover para a direita caso não haja nenhum obstáculo que impeça este movimento.

trepa :: Jogo -> Jogador

A função trepa é a função que faz o jogador trepar caixas e blocos, caso não haja nenhum obstáculo que impeça este movimento.

pegaLargaCaixa :: Jogo -> Jogador

A função pegaLargaCaixa é a função que faz o jogador pegar ou largar caixas, consoante a sua componente tc e se não existir nenhum obstáculo que impeça esta interação.

trocaPecaMapa :: Coordenadas -> Peca -> Mapa -> Mapa

A função trocaPecaMapa é a função que vai trocar a peça p com a peça que está na posição (x,y) do mapa. É útil para as ações de pegar e largar caixas.

trocaPecaLinha :: Int -> Peca -> [Peca] -> [Peca]

A função trocaPecaLinha é a função auxiliar da função trocaPecaMapa que troca a peça p pela peça da posição x de uma linha do mapa.

atualizaMapa :: Mapa -> Jogador -> Movimento -> Mapa

A função atualizaMapa é uma das funções mais importantes desta tarefa pois é responsável por atualizar o mapa m consoante a posição c do jogador e o movimento mov aplicado ao jogador. Caso o movimento seja InterageCaixa atualiza-se a posição da caixa com a qual o jogador interagiu. Qualquer outro movimento vai trocar a peça onde o jogador está pela peça Vazio. Isto deve-se ao inicializar o jogo com umas coordenadas definidas à mão, para que na iteração seguinte não apareça uma peça no lugar de onde o jogador começou o jogo.

mudaPosCaixa :: Jogo -> Movimento -> Mapa

A função mudaPosCaixa vai mudar a posição da caixa com a qual o jogador interage. Se o jogador pegar na caixa em frente a ele, a caixa simplesmente desaparece do mapa, mas é mostrada gráficamente por cima do jogador. Isto deve-se a que a representação gráfica de um jogador com a componente tc igual a True esteja definida sempre a aparecer o carater C por cima dele.

largaCaixa :: Jogo -> Movimento -> Mapa

A função largaCaixa é a função que define onde as coordenadas onde a caixa vai aparecer quando o jogador a largar. Se à frente do jogador estiver uma peça Vazio tem que se verificar se existem mais peças Vazio por baixo da primeira para fazer a caixa "cair" nessa abertura do mapa. Caso contrário, a Caixa é colocada na posição acima da peça que se encontra à frente dele.

atualizaJogo :: Jogo -> Jogo

A função atualizaJogo é uma função importante, que tem o papel de atualizar o Jogo, isto é, baixar o jogador até ao solo caso ele se tenha movido para uma posição com peças Vazio por baixo.

baixaJogador :: Mapa -> Jogador -> Jogador

A função baixaJogador é a função responsável por baixar o jogador para as coordenadas corretas, isto é, para as coordenadas que ficam imediatamente acima do solo.

encontraCoordSolo :: Coordenadas -> Mapa -> Coordenadas

A função encontraCoordSolo é a função que vai calcular ordenada da coluna x onde está declarada a primeira peça considerada como solo, isto é, peças do tipo Bloco e Caixa.

baixaCaixa :: Mapa -> Coordenadas -> Mapa

A função baixaCaixa é a função que vai fazer com que a peça largada pelo jogador ignore as peças Vazio por baixo dela e "caia" na posição imediatamente acima do solo.

moveJogador :: Jogo -> Movimento -> Jogo

A função moveJogador é a função principal desta Tarefa. Tem o objetivo de alterar o jogo tendo em conta o movimento dado. Para tal:

Move, se possível, o jogador segundo esse movimento;

Atualiza o mapa depois de efetuado o movimento;

Atualiza o jogo, caso o jogador se tenha colocado numa posição "flutuante"

Exemplo de utilização:

>>> moveJogador (Jogo [ [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio],
                        [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                        [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                        [Porta, Vazio, Vazio, Vazio, Caixa, Vazio, Bloco],
                        [Bloco, Bloco, Bloco, Bloco, Bloco, Bloco, Bloco]]  (Jogador (6,1) Oeste True) ) AndarEsquerda  

CX P C<X XXXXXXX

correrMovimentos :: Jogo -> [Movimento] -> Jogo

A função correrMovimentos é a generalização da função moveJogador que vai atualizando o jogo j à medida de vai correndo os movimentos presentes na lista dada.

Exemplo de utilização:

>>> correrMovimentos (Jogo [ [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio],
                            [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                            [Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],
                            [Porta, Vazio, Vazio, Vazio, Caixa, Vazio, Bloco],
                            [Bloco, Bloco, Bloco, Bloco, Bloco, Bloco, Bloco]]  
                            (Jogador (6,1) Oeste True) ) 
                            [AndarEsquerda,Trepar,InterageCaixa ] 

< X P CC X XXXXXXX